+++ app/controllers/folio/api/s3_controller.rb
@@ -65,7 +65,7 @@ class Folio::Api::S3Controller < Folio::Api::BaseController
                                 type:,
                                 existing_id: params[:existing_id].try(:to_i),
                                 web_session_id: session.id.public_id,
-                                user_id: try(:current_user).try(:id),
+                                user_id: Folio::Current.user.try(:id),
                                 attributes: Rails.application.config.folio_direct_s3_upload_attributes_for_job_proc.call(self))
         render json: {}
       else