+++ app/controllers/folio/console/api/current_users_controller.rb
@@ -2,7 +2,7 @@
 
 class Folio::Console::Api::CurrentUsersController < Folio::Console::Api::BaseController
   def console_url_ping
-    current_user.update_console_url!(params.require(:url))
+    Folio::Current.user.update_console_url!(params.require(:url))
     head 200
   end
 end