+++ test/dummy/app/lib/application_component.rb
@@ -5,8 +5,6 @@ class ApplicationComponent < Folio::ApplicationComponent
 
   %i[
     cache_key_base
-    current_user
-    user_signed_in?
     current_header_menu
     current_footer_menu
   ].each do |name|